escnet_0907_requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
arel==0.3.0
async-asgi-testclient==1.4.11
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
black==24.8.0
bleach==6.1.0
cacheout==0.14.1
cachetools==5.5.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dataframe_image==0.2.7
dataset-tools==0.1.9
debugpy==1.8.16
defusedxml==0.7.1
distinctipy==1.3.4
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fast_pytorch_kmeans==0.2.2
fastapi==0.109.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.0.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
geojson==3.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.12.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
h2==4.1.0
hf-xet==1.1.5
hpack==4.0.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
inflect==7.4.0
iopath==0.1.9
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.3
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
numerize==0.12
numpy==1.22.4
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
-e git+https://github.com/shimizuryosuke-888/ESC-Net.git@430ebaa369ca82cc4c75d98fa46f49c2f331cd07#egg=open_clip_torch&subdirectory=open_clip
opencv-python==4.12.0.88
orjson==3.10.15
packaging==25.0
pandas==1.5.2
pandocfilters==1.5.1
pathspec==0.12.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
playwright==1.48.0
portalocker==3.0.0
propcache==0.2.0
protobuf==3.20.3
psutil==5.9.8
PTable==0.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pydicom==2.4.4
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.2
PyJWT==2.9.0
Pympler==1.0.1
pynrrd==0.4.3
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rich==14.0.0
rpds-py==0.20.1
rsa==4.9.1
ruff==0.12.2
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
-e git+https://github.com/shimizuryosuke-888/ESC-Net.git@430ebaa369ca82cc4c75d98fa46f49c2f331cd07#egg=segment_anything&subdirectory=segment-anything
semantic-version==2.10.0
shapely==2.0.2
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.35.1
stringcase==1.2.0
supervisely==6.73.418
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.8.3.dev0
tinycss2==1.2.1
titlecase==2.4.1
tomli==2.2.1
tomlkit==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchinfo==1.8.0
torchvision==0.13.1+cu113
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.0
typeguard==4.4.0
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.15
uvicorn==0.33.0
uvloop==0.21.0
varname==0.15.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.6
xmltodict==0.13.0
yacs==0.1.8
yarl==1.15.2
zipp==3.20.2
zstd==1.5.7.2
